requirements.txt
pymongo
streamlit
pillow
streamlit-lottie
bcrypt
flask_jwt_extended